refactor: reset버튼 분리
