+++ components/characterBox/characterBox.module.scss
@@ -2,7 +2,6 @@
   position: relative;
   width: 380px;
   height: 340px;
-  margin-bottom: 20px;
   overflow: hidden;
   border-radius: 20px;
   background-color: $primary1;
@@ -10,7 +9,8 @@
   @include tablet {
     width: 100%;
     max-width: 388px;
-    height: 248px;
+    height: 288px;
+    margin: 0 auto;
   }
   @include mobile {
     width: 100%;
@@ -20,7 +20,5 @@
 }
 
 .img {
-  width: 100%;
-  height: 100%;
   object-fit: cover;
 }
