+++ components/characterBox/characterBox.tsx
@@ -26,7 +26,7 @@ export default function CharacterBox({
           />
           {'face' in selectedItems && selectedItems.face && (
             <Image
-              className={cx('item')}
+              className={cx('img')}
               src={selectedItems.face}
               alt="face item"
               priority={true}
@@ -35,7 +35,7 @@ export default function CharacterBox({
           )}
           {'body' in selectedItems && selectedItems.body && (
             <Image
-              className={cx('item')}
+              className={cx('img')}
               src={selectedItems.body}
               alt="body item"
               fill={true}
@@ -44,7 +44,7 @@ export default function CharacterBox({
           )}
           {'expression' in selectedItems && selectedItems.expression && (
             <Image
-              className={cx('item')}
+              className={cx('img')}
               src={selectedItems.expression}
               alt="expression item"
               fill={true}
@@ -53,7 +53,7 @@ export default function CharacterBox({
           )}
           {'set' in selectedItems && selectedItems.set && (
             <Image
-              className={cx('item')}
+              className={cx('img')}
               src={selectedItems.set}
               alt="set item"
               fill={true}
