+++ components/inventory/inventory.tsx
@@ -1,33 +1,20 @@
 'use client'
 
-import { ReactNode } from 'react'
-
 import { useQuery } from '@tanstack/react-query'
 import classNames from 'classnames/bind'
-import { useRecoilState, useResetRecoilState } from 'recoil'
+import { useRecoilState } from 'recoil'
 
 import CommonBtn from '@/components/commonBtn/commonBtn'
 import { fetchMyProfileRegisterData } from '@/components/makeMyManual/makeMyManual.api'
-import {
-  selectedBodyItemState,
-  selectedCategoryState,
-  selectedExpressionItemState,
-  selectedFaceItemState,
-  selectedSetState,
-} from '@/components/makeMyManual/makeMyManual.atom'
-import ResetBtn from '@/components/resetBtn/resetBtn'
+import { selectedCategoryState } from '@/components/makeMyManual/makeMyManual.atom'
 
 import styles from './inventory.module.scss'
 import { InventoryCategoryBtnList } from './inventoryCategoryBtnList/inventoryCategoryBtnList'
 import { InventoryItemBoxList } from './inventoryItemBoxList/inventoryItemBoxList'
 
 const cx = classNames.bind(styles)
-type InventoryProps = {
-  characterBox?: ReactNode // ReactNode 타입을 직접 사용합니다.
-}
-export default function Inventory({
-  characterBox,
-}: InventoryProps): React.ReactElement {
+
+export default function Inventory({ resetBtn }: { resetBtn: React.ReactNode }) {
   const { data } = useQuery({
     queryKey: ['myprofileRegister'],
     queryFn: fetchMyProfileRegisterData,
@@ -39,30 +26,14 @@ export default function Inventory({
   const selectedCategoryItems =
     data?.itemsData?.characterItems[selectedCategory] || []
 
-  const resetFace = useResetRecoilState(selectedFaceItemState)
-  const resetBody = useResetRecoilState(selectedBodyItemState)
-  const resetExpression = useResetRecoilState(selectedExpressionItemState)
-  const resetSet = useResetRecoilState(selectedSetState)
-
-  const handleResetBtnClick = () => {
-    resetFace()
-    resetBody()
-    resetExpression()
-    resetSet()
-  }
   return (
     <div className={cx('wrap')}>
-      <div>
-        {characterBox}
-        <div className={cx('manuBar')}>
-          <InventoryCategoryBtnList
-            selectedCategory={selectedCategory}
-            setSelectedCategory={setSelectedCategory}
-          />
-          <div className={cx('resetBtn')}>
-            <ResetBtn onClick={handleResetBtnClick} />
-          </div>
-        </div>
+      <div className={cx('content')}>
+        <div className={cx('resetBtn')}>{resetBtn}</div>
+        <InventoryCategoryBtnList
+          selectedCategory={selectedCategory}
+          setSelectedCategory={setSelectedCategory}
+        />
         <InventoryItemBoxList selectedCategoryItems={selectedCategoryItems} />
       </div>
       <div className={cx('btn')}>
