+++ components/inventory/inventoryCategoryBtnList/inventoryCategoryBtnList.module.scss
@@ -1,17 +1,15 @@
 .listWrap {
   display: flex;
-  justify-content: center;
+  justify-content: flex-start;
   width: 100%;
   gap: 24px;
   @include tablet {
     gap: 26px;
-    justify-content: flex-start;
     width: 388px;
     margin: 0 auto;
   }
   @include mobile {
     gap: 14px;
-    justify-content: flex-start;
     width: 280px;
     margin: 0 auto;
   }
