+++ components/inventory/inventoryItemBoxList/inventoryItemBoxList.module.scss
@@ -1,6 +1,5 @@
 .boxList {
   display: flex;
   justify-content: center;
-  margin-top: 32px;
   gap: 14px;
 }
